+++ src/utils.h
@@ -2,6 +2,6 @@
 #pragma once
 
 long long get_time();
-int atoi_m(const char* c);
-void itoa_m(int n, char* buff);
+long long atoi_m(const char* c);
+void itoa_m(long long n, char* buff);
 float mod(float a, int x);
